Improve error selection dropdown with checkbox filtering
