+++ public/locales/en/translation.json
@@ -254,6 +254,12 @@
   "Unassigned Variables": "Unassigned Variables",
   "Dataset": "Dataset",
   "View only row with errors:": "View only row with errors:",
+  "Show all data": "Show all data",
+  "Show only rows with errors": "Show only rows with errors",
+  "Filter rows:": "Filter rows:",
+  "Verification Rules": "Verification Rules",
+  "Schema Preview": "Schema Preview",
+  "Unverified Data": "Unverified Data",
   "Pass Verification": "Pass Verification",
   "Fail Verification": "Fail Verification",
   "Unmatched Attributes": "Unmatched Attributes",
@@ -264,6 +270,7 @@
   "Select Errors": "Select Errors",
   "Entry Codes": "Entry Codes",
   "Data Type": "Data Type",
+  "Format": "Format",
   "To make it easier to upload a schema you can drag and drop your existing dataset that you are creating the schema for. This schema should be in JSON format": "To make it easier to upload a schema you can drag and drop your existing dataset that you are creating the schema for. This schema should be in JSON format",
   "To make it easier to upload a dataset you can drag and drop your existing dataset that you are creating the schema for. This dataset should be in table format as an Excel or .csv format": "To make it easier to upload a dataset you can drag and drop your existing dataset that you are creating the schema for. This dataset should be in table format as an Excel or .csv format",
   "To make it easier to create a schema you can drag and drop your existing dataset that...": "To make it easier to create a schema you can drag and drop your existing dataset that you are creating the schema for. This dataset should be in table format as an Excel or .csv format. The first row must contain the column headers (aka attribute names). \n \n Note that no data leaves your device, and no data is uploaded to a server. All the processing to copy the column headers (attribute names) into the draft schema happens on your device.",
