+++ public/locales/fr/translation.json
@@ -253,9 +253,15 @@
   "Unassigned Variable": "Variable non attribuée",
   "Dataset": "Base de données",
   "View only row with errors:": "Afficher uniquement la ligne contenant des erreurs:",
+  "Show all data": "Afficher toutes les données",
+  "Show only rows with errors": "Afficher uniquement les lignes avec des erreurs",
+  "Filter rows:": "Filtrer les lignes:",
+  "Verification Rules": "Règles de vérification",
+  "Schema Preview": "Aperçu du schéma",
+  "Unverified Data": "Données non vérifiées",
   "Pass Verification": "Vérification de réussite",
   "Fail Verification": "Échec de la vérification",
-  "Unmatched Attributes": "Entrez du texte",
+  "Unmatched Attributes": "Des attributs inégalés",
   "Export Verified Data": "Exporter des données vérifiées",
   "Keep original data column headers": "Conserver les en-têtes de colonnes de données d'origine",
   "Change to Schema column headers": "Modification des en-têtes de colonne de schéma",
@@ -263,6 +269,7 @@
   "Select Errors": "Sélectionnez les erreurs",
   "Entry Codes": "Codes d'entrée",
   "Data Type": "Type de données",
+  "Format": "Format",
   "To make it easier to upload a schema you can drag and drop your existing dataset that you are creating the schema for. This schema should be in JSON format": "Pour faciliter le téléchargement d'un schéma, vous pouvez faire glisser et déposer votre ensemble de données existant pour lequel vous créez le schéma. Ce schéma doit être au format JSON",
   "To make it easier to upload a dataset you can drag and drop your existing dataset that you are creating the schema for. This dataset should be in table format as an Excel or .csv format": "Pour faciliter le téléchargement d'un ensemble de données, vous pouvez glisser-déposer l'ensemble de données existant pour lequel vous créez le schéma. Cet ensemble de données doit être sous forme de tableau au format Excel ou .csv",
   "To make it easier to create a schema you can drag and drop your existing dataset that...": "Pour faciliter la création d'un schéma, vous pouvez faire glisser et déposer votre ensemble de données existant pour lequel vous créez le schéma. Cet ensemble de données doit être sous forme de tableau au format Excel ou .csv. La première ligne doit contenir les en-têtes de colonnes (c'est-à-dire les noms d'attributs). \n \n Notez qu'aucune donnée ne quitte votre appareil et qu'aucune donnée n'est téléchargée sur un serveur. Tout le traitement pour copier les en-têtes de colonnes (noms d'attributs) dans le brouillon de schéma se produit sur votre appareil.",
